requirements.txt
Flask==2.3.3
Flask-Session==0.5.0
Flask-Caching==2.0.2
Flask-WTF==1.1.1
requests==2.31.0
email_validator==2.0.0.post2
python-dotenv==1.0.0
aiohttp==3.8.5
asyncio==3.4.3